+++ Magento/Mtf/TestCase/Injectable.php
@@ -69,6 +69,13 @@ abstract class Injectable extends Functional
     protected $currentVariation = [];
 
     /**
+     * Is incomplete test.
+     *
+     * @var bool
+     */
+    protected $isIncomplete = false;
+
+    /**
      * Constructs a test case with the given name
      *
      * @constructor
@@ -300,5 +307,26 @@ abstract class Injectable extends Functional
             'Magento\Mtf\Constraint\Composite',
             ['codeConstraints' => array_keys($constraintsArray)]
         );
+    }
+
+    /**
+     * Get incomplete status of test.
+     *
+     * @return bool
+     */
+    public function getIsIncomplete()
+    {
+        return $this->isIncomplete;
+    }
+
+    /**
+     * Set incomplete status of test.
+     *
+     * @param bool $status
+     * @return void
+     */
+    public function setIsIncomplete($status)
+    {
+        $this->isIncomplete = $status;
     }
 }